Removed the remote debugging functions.
I used to find these quite useful for helping people debug
configuration problems. I don't think they present a security
problem, but I can see there is the possibility that it can
be PERCEIVED as an issue, and I don't want to be responsible
for that, nor blamed for any break-ins (warranty or not).

So I'll find another way to do remote debugging.