Added some bug-fixing code from Eloy Lafuente that fixes a quiz problem.
What happens is that adding random questions used to set the timestamps
of these questions to the same value.  This is fine in normal operation,
but it stuffs up the backup of that course, overwriting the question.
When the course is restored, the missing questions cause quizzes to
have lots of red notices like "Question not found".

Anyhow, this patch should safely fix older installations that may have
these duplicate questions lying around.

THANKS ELOY!