This is in principle the theme I've shown in the forums. I changed the colors and made it XHTML 1.0 suitable. It relies on the new DIV framework I added to the course, forum and the wiki pages. The layout-table with named cells is used for the course pages.
For an overvies of the DIV framework and the DIVs in the header and footer please look at http://moodle.org/mod/wiki/view.php?id=2935&userid=0&groupid=0&wikipage=CoursePage

The CSS is not yet well structured - it's just working for a starting. For the offical release I'll have to rebuild the CSS to match the final needs. It's lacking an easy CSS for beginners too.

Please tell me, if you find some bugs or have some remarks - or if you just like it ;-)

Thank you
Urs