New feature for resources ... Local File Resources
A Resource Parameter controls this (it defaults to OFF).

When on, you see a third button when selecting a file, which lets
you select a LOCAL file on your own filesystem (but it isn't uploaded).

The local file is included in the web page when viewed.  All users see
a link that helps them set up a user_preference for the location of
these files on their own system (usually a CD_ROM drive).