Merging from STABLE:
Fix for bug 2925:

When editing is turned on, the Main Menu should always display and give
the option of adding resources/activities in section 0 of the site.

Bad optimization code caused the drop downs to not be shown if there
was no section 0 (i.e., the site never had a news forum or an activity
in the Main Menu block).

FINALLY! :)

PS: Version bump, so that we know if people have the latest version of
the block. I 'm going to bump versions from now on when bugs are fixed.