Merging from HEAD:
Fix for bug 3981 (also quite lively discussed at
http://moodle.org/mod/forum/discuss.php?d=29564):

When the user was not logged in and was not a guest, DST compensation was not
being done. I guess deciding to store the computationally expensive DST cache
in $USER was a major oops on my part. I 've moved everything to $SESSION and
now it works as it should.