NEW: Progress bar option.  Now a progress bar can be optionally displayed at the bottom of the lesson page. NEW: Display Left Menu IF.  The left menu will only display if the student has a grade better than the one set in the lesson settings.  This is nice to allow only students who are reviewing the lesson to have access to the left menu.
BUG FIX: In Windows IE, the content was being aligned center.  Now, the content is explicitly aligned to left as default.
BUG FIX: High Scores variable was not spelled correctly at the End of Lesson.

ADDED: Div tags around branch tables.  This will allow users to choose to display them or not or move them around

CHANGED: converted more of the links to the stylized button.  Also, removed a lot of forms and javascript and replaced it with links with the variables in the href.
CHANGED: the slide show styles have been moved to the style sheet.
CHANGED: removed the CDC comments

Other minor changes have been made to the code as well.