Comments from Andy: config.php contains a new option,  $THEME->chameleonteachereditenabled  by default this is false, preventing teachers from editing chameleon course themes.  css.php has been changed a little to make use of this new property, it now also checks $CFG->allowcoursethemes before letting a teacher edit anything.
There have been some bug fixes in the handling of border shorthands.

Some workarounds to avoid a pretty big IE bug - the addRule method in IE's version of DOM 2 CSS  (which I use to create the live preview of changes), can crash the browser if the selector being added contains certain characters. Now as chameleon generates the preview it checks the selectors against a list of allowed characters - if the selector contains any character not on the list it will be ignored.

I don't think I'm quite there yet, but I made a bit of a change in some of the event handling code which means support for IE 5.5 seems to be getting closer :-)