Fix for Bug #4910 - stringnames with - . and / don't show up in translation editor
Modified /admin/lang.php to allow us to still use periods in language string keys. The lang.php script just replaces the periods with ##46# for the form inputs and then puts the periods back in before saving to the language file.

This means that there is no need to update language files at all. They will just work even with periods in the string keys.