altered require statement for config.
Not sure why, but the referential file path was causing cron to fail because config couldn't be found when running from scheduled cron (running in browser was fine!

It looks tidier anyway to have both require's done the same way, so now both user $dirroot to locate files required and this works with scheduled cron.