formslib: MDL-13210 - partial - try to make validation errores less annoying.
I have made them float right, and reduced the padding on the red box, so they don't cuause the height of the form to change so much, if the error fits to the right of the form control. I realise that some people may not like this, which I why I have done this in head only for now. I will revert this if told to.