debug / lang string: MDL-19467 Debugging option to show where each lang string came from.
New option on Administration -> Server -> Debugging called
'Show origin of languages strings' ($CFG->debugstringids).

If this is set, get_string will add the name of the lang file and the id of the string
to each string it gets, so instead of returning 'No', it will return '{No admin/checkboxno}'.