MDL-19932 Rubric grading strategy implemented
The only weak point here is that we store a raw grade into
workshop_grades and not a direct id of the selected level. Therefore,
when re-assessing, we need to actually guess what level the assessor
chose previously. This is not problem if there are not two levels with
the same grade. Such case is not common when using Rubric. In the
future, this may get refactored so Rubric would use its own storage of
filled assessment forms.