accesslib: MDL-22089 Unassigning role from one user causes warning to happen
The 3rd parameter to role_unassign() was renamed to $groupid_ignored some weeks
ago (as part of MDL-21655), but the $args array that is used to build the query
was not updated to reflect that change. Simply remove 'groupid' from the $args
array so it isn't used in the query at all.