MDL-21695 gradebook: Extra credit help strings
String aggregationcoefextraweight is used when editing an item which is
aggregated by Mean of grades (with extra credit) and was reworded to
Extra credit weight (because in fact it is the weight of the extra
credit - if it equals zero, the grade is not extra)

String aggregationcoefextrasum is used for Sum of grades and Simple
weighted mean of grades, meaning a checkbox extra credit yes/no

String aggregationcoefextra is used in the header of table Categories
and items (either Simple view or Full view). It must contain the information
about both meanings of aggregation.

AMOS BEGIN
 HLP grade/aggregationcoefcombo.html,[aggregationcoefextra_help,core_grades]
 HLP grade/aggregationcoefextra.html,[aggregationcoefextraweight_help,core_grades]
 HLP grade/aggregationcoefextrasum.html,[aggregationcoefextrasum_help,core_grades]
AMOS END