Portfolio plugins explicitly requires parent library with superclass definition
I have discovered a bug with GoogleDocs portfolio in events. An event
handler was defined in googledocs/lib.php and so the cron included just
this library. But the inclusion failed because the superclass
portfolio_plugin_push_base was not known to cron.

IMHO it is generally good habit to require_once the file which defines
the parent classes.