MDL-12192 Language specific label separator instead of hardcoded ': '
New string is available. Use get_string('labelsep', 'langconfig') to be used
as a separator between a label and the labelled text. If you want to
produce <label> HTML element, you may be interested in the new
html_writer::label() method. Simple selects and URL selects now use it.

There are bambillion places where ': ' is hardcoded as a separator. I do
not think we have to fix them manually now (and mass search and replace
would be tricky) but everybody should stop adding new.

Do not abuse 'labelsep' for any other purposes. Note that it may not be
colon in some languages or sites.