MDL-15976 Shortening 'Taiwan (Province of China)' to 'Taiwan'.
Our rationale before was to stick to the ISO-3166 standard no matter what, since we don't want to get involved in territorial issues.  However, in this case no information is lost and it would make a lot of people happier (and no one unhappy, I think), so I'm going to make an exception.  Hope it's not a can of worms.