MDL-21861 Made these assignment events more generic, so that they can be used by other activity modules (like Workshop etc).  assessable_file_uploaded and assessable_files_done
We are doing it this way because MULTIPLE plagiarism **plugins** may need to handle these events using cron, so it's good to get them in a queue.

The other way to do this would be to call a function in plagiarism core and have that generate the event, but there is not really any advantage in that.   Conceivably something other than plagiarism plugins might be interested in these events later on.