MDL-21695 Cleaning up old help help files after the migration to strings was done
The legacy help files could remain here if I forgot to remove them in
the migration commit or if they were not supposed to be migrated into a
string. There is no setHelpButton() in Moodle code now. The files could
be referenced by $OUTPUT->old_help_icon() yet but that should be fixed
soon.