MDL-24896 Do not access PHPMailer's private properties but use public API only
There was a problem with properties $to, $ReplyTo and $error_count that
are now private. $error_count is accessible via public IsError() method.
For the other two, we needed them just because of post-processing charset
conversion. So I just made sure that the charset is converted before
the recipent and reply-to names are added to the PHPMailer.

Also, I do not instantiate Typo3 during every email if it is not needed
and I pass the character set to it in lower case now as stated in
textlib documentation.